Various tweaks to DDG types
