+++ packages/jaeger-ui/src/model/ddg/getDdgModelKey.tsx
@@ -1,0 +1,19 @@
+// Copyright (c) 2019 Uber Technologies, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import { TDdgModelParams } from './types';
+
+export default function getDdgModelKey({ service, operation = '*', start, end }: TDdgModelParams) {
+  return `${service}\n${operation}\n${start}\n${end}`;
+}
