+++ packages/jaeger-ui/src/reducers/deep-dependency-graph.tsx
@@ -14,13 +14,13 @@
 
 import { handleActions } from 'redux-actions';
 
-import { actionTypes } from '../actions/deep-dependency-graph';
+import { actionTypes } from '../actions/ddg';
 import { fetchDeepDependencyGraph } from '../actions/jaeger-api';
 import { fetchedState } from '../constants';
 import { ApiError } from '../types/api-error';
+import getDdgModelKey from '../model/ddg/getDdgModelKey';
 import transformDdgData from '../model/ddg/transformDdgData';
 import {
-  stateKey,
   EViewModifier,
   TDdgActionMeta,
   TDdgAddViewModifierPayload,
@@ -28,15 +28,14 @@ import {
   TDdgPayload,
   TDdgRemoveViewModifierFromIndicesPayload,
   TDdgRemoveViewModifierPayload,
-  TDdgState,
-  TDdgStateEntry,
   TDdgViewModifierRemovalPayload,
 } from '../model/ddg/types';
+import TDdgState, { TDdgStateEntry } from '../types/TDdgState';
 import guardReducer, { guardReducerWithMeta } from '../utils/guardReducer';
 
 export function addViewModifier(state: TDdgState, { payload }: { payload: TDdgAddViewModifierPayload }) {
   const { visibilityIndices, viewModifier } = payload;
-  const key = stateKey(payload);
+  const key = getDdgModelKey(payload);
   const stateEntry: TDdgStateEntry | void = state[key];
   if (!stateEntry || stateEntry.state !== fetchedState.DONE) {
     console.warn('Cannot set view modifiers for unloaded Deep Dependency Graph'); // eslint-disable-line no-console
@@ -62,7 +61,7 @@ export function viewModifierRemoval(
   { payload }: { payload: TDdgViewModifierRemovalPayload }
 ) {
   const { visibilityIndices, viewModifier } = payload;
-  const key = stateKey(payload);
+  const key = getDdgModelKey(payload);
   const stateEntry: TDdgStateEntry | void = state[key];
   if (!stateEntry || stateEntry.state !== fetchedState.DONE) {
     console.warn('Cannot change view modifiers for unloaded Deep Dependency Graph'); // eslint-disable-line no-console
@@ -95,7 +94,7 @@ export function viewModifierRemoval(
 
 export function fetchDeepDependencyGraphStarted(state: TDdgState, { meta }: { meta: TDdgActionMeta }) {
   const { query } = meta;
-  const key = stateKey(query);
+  const key = getDdgModelKey(query);
   return {
     ...state,
     [key]: {
@@ -110,7 +109,7 @@ export function fetchDeepDependencyGraphDone(
 ) {
   const { query } = meta;
   const { service, operation } = query;
-  const key = stateKey(query);
+  const key = getDdgModelKey(query);
   return {
     ...state,
     [key]: {
@@ -126,7 +125,7 @@ export function fetchDeepDependencyGraphErred(
   { meta, payload }: { meta: TDdgActionMeta; payload: ApiError }
 ) {
   const { query } = meta;
-  const key = stateKey(query);
+  const key = getDdgModelKey(query);
   return {
     ...state,
     [key]: {
